address Copilot review: request-id collision, poisoned bootstraps, key perms
All eight findings accepted — several are real bugs, two of them in code added
in the previous review round.

myotis-net:
- in_flight is now keyed by (protocol, InboundRequestId). The id is issued PER
  request_response::Behaviour and this host runs seven of them, so two protocols
  can each mint id 1: the second insert overwrote the first while both lengths
  stayed in the total, and the first terminal event released the wrong entry.
  Repeated collisions leak the budget until every response is refused — i.e. the
  bound added last round could disable the responder it was meant to protect.
- parse_request_ssz_expecting compares the declared length as u64 BEFORE
  narrowing. On a 32-bit target (android armv7) `as usize` truncates, so
  2^32 + expected passed the equality check and was accepted as a short request.

roost:
- Bootstrap misses drain FIFO (VecDeque + membership set) instead of BTreeSet
  key order, which let a caller starve a legitimate checkpoint indefinitely by
  enqueueing lexicographically smaller roots before each tick.
- Transient upstream failures are requeued instead of negative-cached. The root
  is marked attempted BEFORE the fetch, so one Nimbus timeout or 5xx
  permanently poisoned a wallet's pinned checkpoint — every later request
  refused with no re-fetch. Only 4xx is definitive now; FetchError carries the
  distinction.
- The identity key is created with create_new + mode 0600 in one call. Plain
  fs::write used the process umask (the existing dev key is 664, confirming it),
  so any local user could copy it and impersonate the server; exists()-then-
  write() was also a race where a second start could clobber a published
  identity. Keys written by older builds now get a loud warning.
- The archive no longer truncates on a corrupted length in a NON-tail record. A
  bad length runs past EOF exactly like an interrupted write, and treating it as
  a torn tail deleted every valid record after it — the opposite of the salvage
  guarantee this format exists for. It now checks whether an intact record
  follows before classifying, and candidate offsets are parsed and checksummed
  rather than accepted on a REC_MAGIC byte match.
- REST bodies are streamed under MAX_BODY_BYTES (32 MiB) with the declared
  Content-Length refused up front, so a malfunctioning upstream cannot exhaust
  the daemon before split_updates gets a chance to bound anything.
- --port rejects 0, which the error message already promised: it asks for an
  ephemeral port and the daemon would then advertise an undialable /tcp/0.

Verified: new keys are 0600 and an old 0644 key warns; --port 0 is rejected;
wallet still reaches SYNCED from roost alone. roost 33 tests + clippy clean,
myotis-net 212 tests.

+++ rust/myotis-net/src/codec.rs
@@ -167,7 +167,11 @@ pub fn parse_request_ssz_expecting(raw: &[u8], expected: Option<usize>) -> Optio
         return None;
     }
     if let Some(expected) = expected {
-        if uncompressed_len as usize != expected {
+        // Compare as u64 BEFORE narrowing. `uncompressed_len as usize` would
+        // truncate on a 32-bit target (android armv7 is one), so a declared
+        // 2^32 + expected would pass this check and then be treated as a short
+        // request. Equality here makes the conversion below provably safe.
+        if uncompressed_len != expected as u64 {
             return None;
         }
     }
